We only have one DomParser
