+++ src/commonMain/kotlin/com/ashampoo/xmp/impl/XMPMetaParser.kt
@@ -47,7 +47,7 @@ internal object XMPMetaParser {
 
         val actualOptions = options ?: ParseOptions()
 
-        val document = XmlUtilDomParser.parseDocumentFromString(input)
+        val document = DomParser.parseDocumentFromString(input)
 
         val xmpmetaRequired = actualOptions.getRequireXMPMeta()
